# agent_builder/openapi/tour_data_api.yaml
openapi: 3.0.0
info:
  title: Tour Intelligence Data API
  description: >
    Deterministic, parameterized lookups against the tour_intelligence
    BigQuery dataset, plus the enthusiasm-scoring, city-ranking, and
    grounding-check SDK tools. Used in place of a native Data Store tool
    because these are exact-key structured lookups, not semantic search.
  version: "1.0.0"
servers:
  - url: https://tour-data-api-602700957663.us-central1.run.app
paths:
...
  /campaign_strategy_chat:
    post:
      operationId: campaignStrategyChat
      summary: >
        Conversational pre-fill assistant for the New Campaign form. Not a
        Playbook tool used by any of the 5 MVP agents -- called directly by
        the dashboard BFF. Purely advisory: returns a suggested_campaign the
        user reviews and submits through the existing createCampaign path;
        this route never writes anything itself.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role: { type: string }
                      content: { type: string }
                strategy_text: { type: string, nullable: true }
      responses:
        "200":
          description: Assistant reply, plus a suggested campaign once enough information is gathered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply: { type: string }
                  ready: { type: boolean }
                  suggested_campaign:
                    type: object
                    nullable: true
                    properties:
                      title: { type: string }
                      campaign_type: { type: string }
                      genre: { type: string }
                      talent_roster: { type: array, items: { type: string } }
                      stops:
                        type: array
                        items:
                          type: object
                          properties:
                            city_id: { type: string }
                            stop_date: { type: string }
        "502":
          description: Gemini call failed.
...
  /campaigns:
    get:
      operationId: getCampaign
      summary: Campaign-level metadata — genre, campaign type, and talent roster — needed by the Fan Enthusiasm Agent.
      parameters:
        - name: campaign_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign record.
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaign_id: { type: string }
                  title: { type: string }
                  campaign_type: { type: string }
                  genre: { type: string }
                  talent_roster: { type: array, items: { type: string } }
                  status: { type: string }
        "404":
          description: No matching record for the given campaign_id.
    post:
      operationId: createCampaign
      summary: Create a new campaign. Used by the dashboard's "New Campaign" flow — not currently called by any Playbook.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [campaign_id, title, campaign_type, genre]
              properties:
                campaign_id: { type: string }
                title: { type: string }
                campaign_type: { type: string }
                genre: { type: string }
                talent_roster: { type: array, items: { type: string } }
                status: { type: string }
      responses:
        "200":
          description: Insert confirmation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaign_id: { type: string }
                  status: { type: string }
        "500":
          description: Insert failed.
...
  /campaign_stops:
    get:
      operationId: getCampaignStops
      summary: Ordered list of city stops for a campaign, joined with city names.
      parameters:
        - name: campaign_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign stops in sequence order.
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaign_id: { type: string }
                  stops:
                    type: array
                    items:
                      type: object
                      properties:
                        city_id: { type: string }
                        city_name: { type: string }
                        stop_date: { type: string, format: date }
                        sequence_order: { type: integer }
                        event_format: { type: string, nullable: true }
    post:
      operationId: createCampaignStops
      summary: >
        Add city stops to a campaign. Used by the dashboard's "New Campaign"
        flow — not currently called by any Playbook. city_id is restricted to
        the cities with full seeded coverage (mumbai, london, tokyo,
        sao_paulo, new_york) since fan_signals has no live-search fallback.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [campaign_id, stops]
              properties:
                campaign_id: { type: string }
                stops:
                  type: array
                  items:
                    type: object
                    required: [city_id]
                    properties:
                      city_id: { type: string }
                      stop_date: { type: string, format: date }
                      sequence_order: { type: integer }
                      event_format: { type: string, nullable: true }
      responses:
        "200":
          description: Insert confirmation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaign_id: { type: string }
                  status: { type: string }
                  count: { type: integer }
        "400":
          description: Missing fields or an unsupported city_id.
        "500":
          description: Insert failed.
  /city_briefs:
    post:
      operationId: insertCityBrief
      summary: Persist a finalized city brief as a real BigQuery row.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [brief_id, campaign_id, city_id, status]
              properties:
                brief_id: { type: string }
                campaign_id: { type: string }
                city_id: { type: string }
                generated_at: { type: string, format: date-time }
                status: { type: string }
                enthusiasm_score: { type: number }
                culture_summary: { type: string }
                local_delight_summary: { type: string }
                talent_brief_json: { type: string }
                grounding_check_passed: { type: boolean }
                grounding_check_notes: { type: string }
                delight_card_url: { type: string }
      responses:
        "200":
          description: Insert confirmation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  brief_id: { type: string }
                  status: { type: string }
        "500":
          description: Insert failed.
...
  /rank_cities:
    post:
      operationId: rankCities
      summary: Rank a campaign's cities by strategic value (tier, then enthusiasm score) for the Compare Cities screen.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [city_records]
              properties:
                city_records:
                  type: array
                  items:
                    type: object
                    properties:
                      city_id: { type: string }
                      enthusiasm_score: { type: number }
                      city_importance_tier: { type: string }
      responses:
        "200":
          description: Ranked list with strategic_rank assigned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ranked:
                    type: array
                    items:
                      type: object
...